+++ karma.conf.js
@@ -60,8 +60,9 @@ module.exports = (config) => {
         files: [
             srcCssRecursivePath,
             srcRecursivePath,
+            'node_modules/jquery/dist/jquery.min.js',
             'node_modules/powerbi-visuals-utils-testutils/lib/index.js',
-            'node_modules/jasmine-jquery/lib/jasmine-jquery.js',
+            'node_modules/jasmine-jquery/lib/jasmine-jquery.js',            
             recursivePathToTests,
             {
                 pattern: srcOriginalRecursivePath,