+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+## 1.4.9
+* Added new formatting option for Category labels "Maximum width" to adjust label width for horizontal oriented bullets.
+* Horizontal oriented bullets now renders closer to each other when axis is off.
+* Fixed axis tick labels overlapping.
+* Visual is updated to API 1.7.0
+
 ## 1.4.8
 * Bars' values respect selection of other visuals now.